refactor: ViewファイルとFeatureファイルに分離
